requirements.txt
click==6.7
PyYAML==3.12
Jinja2==2.10

git+https://github.com/cedadev/cc-yaml
